backend/requirements.txt
# NDA Guardrail Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

# AI & Embeddings
groq==0.11.0
sentence-transformers==3.1.1

# Database
supabase==2.9.1
